refactored stuff
